requirements.txt
langchain
langchain_core
langchain_community
PyPDF2


python-dotenv
langchain_openai
langchain_google_genai
langchain_groq


sentence_transformers
chromadb